# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/group": {
      "post": {
        "operationId": "GroupController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Group"
        ]
      },
      "get": {
        "operationId": "GroupController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Group"
        ]
      }
    },
    "/api/group/{id}": {
      "get": {
        "operationId": "GroupController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Group"
        ]
      },
      "patch": {
        "operationId": "GroupController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Group"
        ]
      },
      "delete": {
        "operationId": "GroupController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Group"
        ]
      }
    },
...
    "/api/region": {
      "post": {
        "operationId": "RegionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      },
      "get": {
        "operationId": "RegionController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      }
    },
    "/api/region/{id}": {
      "get": {
        "operationId": "RegionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      },
      "patch": {
        "operationId": "RegionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRegionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      },
      "delete": {
        "operationId": "RegionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      }
    },
...
    "/api/payment": {
      "post": {
        "operationId": "PaymentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Payment"
        ]
      },
      "get": {
        "operationId": "PaymentController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payment"
        ]
      }
    },
    "/api/payment/{id}": {
      "get": {
        "operationId": "PaymentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payment"
        ]
      },
      "patch": {
        "operationId": "PaymentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payment"
        ]
      },
      "delete": {
        "operationId": "PaymentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payment"
        ]
      }
    },
...
    "/api/group-teacher": {
      "post": {
        "operationId": "GroupTeacherController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupTeacherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "GroupTeacher"
        ]
      },
      "get": {
        "operationId": "GroupTeacherController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupTeacher"
        ]
      }
    },
    "/api/group-teacher/{id}": {
      "get": {
        "operationId": "GroupTeacherController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupTeacher"
        ]
      },
      "patch": {
        "operationId": "GroupTeacherController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupTeacherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupTeacher"
        ]
      },
      "delete": {
        "operationId": "GroupTeacherController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupTeacher"
        ]
      }
    },
...
    "/api/lesson": {
      "post": {
        "operationId": "LessonController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      },
      "get": {
        "operationId": "LessonController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
    "/api/lesson/{id}": {
      "get": {
        "operationId": "LessonController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      },
      "patch": {
        "operationId": "LessonController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      },
      "delete": {
        "operationId": "LessonController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
...
    "/api/homework-result": {
      "post": {
        "operationId": "HomeworkResultController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHomeworkResultDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "HomeworkResult"
        ]
      },
      "get": {
        "operationId": "HomeworkResultController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HomeworkResult"
        ]
      }
    },
    "/api/homework-result/{id}": {
      "get": {
        "operationId": "HomeworkResultController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HomeworkResult"
        ]
      },
      "patch": {
        "operationId": "HomeworkResultController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHomeworkResultDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HomeworkResult"
        ]
      },
      "delete": {
        "operationId": "HomeworkResultController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HomeworkResult"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "N15",
            "description": "Group name"
          },
          "course_id": {
            "type": "number",
            "example": 1,
            "description": "Course ID"
          },
          "lesson_days": {
            "example": [
              1,
              2,
              3
            ],
            "description": "Lesson days (1-7)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lesson_start_time": {
            "type": "string",
            "example": "12:00",
            "description": "Lesson start time"
          },
          "lesson_end_time": {
            "type": "string",
            "example": "13:00",
            "description": "Lesson end time"
          },
          "start_date": {
            "type": "string",
            "example": "2025-11-11",
            "description": "Start date (YYYY-MM-DD)"
          },
          "room": {
            "type": "string",
            "example": "Room name",
            "description": "Room name or number"
          },
          "floor": {
            "type": "number",
            "example": 2,
            "description": "Floor number"
          },
          "branch_id": {
            "type": "number",
            "example": 1,
            "description": "Branch ID"
          },
          "is_active": {
            "type": "boolean",
            "example": true,
            "description": "Is group active"
          }
        },
        "required": [
          "name",
          "course_id",
          "lesson_days",
          "lesson_start_time",
          "lesson_end_time",
          "start_date",
          "room",
          "floor",
          "branch_id",
          "is_active"
        ]
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateRegionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the region",
            "example": "Toshkent"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateRegionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the region",
            "example": "Toshkent"
          }
        }
      },
...
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "number",
            "description": "The id of the student",
            "example": 1
          },
          "payment_type": {
            "type": "string",
            "description": "The payment type",
            "example": "monthly"
          },
          "amount": {
            "type": "number",
            "description": "The price of the payment",
            "example": 100000
          },
          "payment_method_id": {
            "type": "number",
            "description": "The payment method id",
            "example": 1
          },
          "transaction_id": {
            "type": "string",
            "description": "The transaction id",
            "example": "1234567890"
          }
        },
        "required": [
          "user_id",
          "payment_type",
          "amount",
          "payment_method_id",
          "transaction_id"
        ]
      },
      "UpdatePaymentDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "number",
            "description": "The id of the student",
            "example": 1
          },
          "payment_type": {
            "type": "string",
            "description": "The payment type",
            "example": "monthly"
          },
          "amount": {
            "type": "number",
            "description": "The price of the payment",
            "example": 100000
          },
          "payment_method_id": {
            "type": "number",
            "description": "The payment method id",
            "example": 1
          },
          "transaction_id": {
            "type": "string",
            "description": "The transaction id",
            "example": "1234567890"
          }
        }
      },
...
      "CreateGroupTeacherDto": {
        "type": "object",
        "properties": {
          "teacher_id": {
            "type": "number",
            "example": 1,
            "description": "Teacher ID"
          },
          "group_id": {
            "type": "number",
            "example": 1,
            "description": "Group ID"
          },
          "is_main": {
            "type": "boolean",
            "example": true,
            "description": "Is main teacher"
          },
          "duration": {
            "type": "number",
            "example": 10,
            "description": "Duration in weeks/months"
          },
          "start_month": {
            "type": "number",
            "example": 1,
            "description": "Start month (1-12)"
          },
          "start_time": {
            "type": "string",
            "example": "12:00",
            "description": "Start time"
          },
          "end_time": {
            "type": "string",
            "example": "13:00",
            "description": "End time"
          }
        },
        "required": [
          "teacher_id",
          "group_id",
          "is_main",
          "duration",
          "start_month",
          "start_time",
          "end_time"
        ]
      },
      "UpdateGroupTeacherDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateLessonDto": {
        "type": "object",
        "properties": {
          "group_id": {
            "type": "number",
            "description": "The id of the group",
            "example": 1
          },
          "teacher_id": {
            "type": "number",
            "description": "The id of the teacher",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "The title of the lesson",
            "example": "Lesson 1"
          },
          "homework": {
            "type": "string",
            "description": "The homework of the lesson",
            "example": "Homework 1"
          },
          "xp": {
            "type": "number",
            "description": "The xp of the lesson",
            "example": 100
          },
          "is_exam": {
            "type": "boolean",
            "description": "The is_exam of the lesson",
            "example": true
          },
          "start_time": {
            "type": "string",
            "description": "The start_time of the lesson",
            "example": "12:00"
          },
          "end_time": {
            "type": "string",
            "description": "The end_time of the lesson",
            "example": "13:00"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Lesson date",
            "example": "2023-09-01"
          },
          "description": {
            "type": "string",
            "description": "Lesson description",
            "example": "First lesson"
          }
        },
        "required": [
          "group_id",
          "teacher_id",
          "title",
          "homework",
          "xp",
          "is_exam",
          "start_time",
          "end_time",
          "description"
        ]
      },
      "UpdateLessonDto": {
        "type": "object",
        "properties": {
          "group_id": {
            "type": "number",
            "description": "The id of the group",
            "example": 1
          },
          "teacher_id": {
            "type": "number",
            "description": "The id of the teacher",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "The title of the lesson",
            "example": "Lesson 1"
          },
          "homework": {
            "type": "string",
            "description": "The homework of the lesson",
            "example": "Homework 1"
          },
          "xp": {
            "type": "number",
            "description": "The xp of the lesson",
            "example": 100
          },
          "is_exam": {
            "type": "boolean",
            "description": "The is_exam of the lesson",
            "example": true
          },
          "start_time": {
            "type": "string",
            "description": "The start_time of the lesson",
            "example": "12:00"
          },
          "end_time": {
            "type": "string",
            "description": "The end_time of the lesson",
            "example": "13:00"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Lesson date",
            "example": "2023-09-01"
          },
          "description": {
            "type": "string",
            "description": "Lesson description",
            "example": "First lesson"
          }
        }
      },
...
      "CreateHomeworkResultDto": {
        "type": "object",
        "properties": {
          "student_homework_id": {
            "type": "number",
            "description": "Student homework ID",
            "example": 1
          },
          "teacher_id": {
            "type": "number",
            "description": "Teacher ID",
            "example": 1
          },
          "ball": {
            "type": "number",
            "description": "Ball (0-100)",
            "example": 85
          },
          "comment": {
            "type": "string",
            "description": "Comment",
            "example": "Good work, but needs improvement in..."
          }
        },
        "required": [
          "student_homework_id",
          "teacher_id",
          "ball",
          "comment"
        ]
      },
      "UpdateHomeworkResultDto": {
        "type": "object",
        "properties": {
          "student_homework_id": {
            "type": "number",
            "description": "Student homework ID",
            "example": 1
          },
          "teacher_id": {
            "type": "number",
            "description": "Teacher ID",
            "example": 1
          },
          "ball": {
            "type": "number",
            "description": "Ball (0-100)",
            "example": 85
          },
          "comment": {
            "type": "string",
            "description": "Comment",
            "example": "Good work, but needs improvement in..."
          }
        }
      },
...
    }
  }
}